+++ readium/navigator/src/main/java/org/readium/r2/navigator/Navigator.kt
@@ -72,7 +72,16 @@ interface Navigator {
      */
     fun goBackward(animated: Boolean = false, completion: () -> Unit = {}): Boolean
 
-    interface Listener
+    interface Listener {
+
+        /**
+         * Called when the navigator will interrupt the linear reading progression to jump to the
+         * given locator.
+         *
+         * This occurs for example when clicking on internal links or calling [go] programmatically.
+         */
+        fun onJumpToLocator(locator: Locator) {}
+    }
 
     @Deprecated("Use [currentLocator.value] instead", ReplaceWith("currentLocator.value"))
     val currentLocation: Locator? get() = currentLocator.value
@@ -194,4 +203,6 @@ interface MediaNavigator : Navigator {
      * Seeks relatively from the current position in the current resource.
      */
     fun seekRelative(offset: Duration)
+
+    interface Listener : Navigator.Listener
 }